+++ src/context.ts
@@ -2,7 +2,6 @@ import { type ZodError } from "zod"
 import { isSupportedBodyMethod } from "./assert.js"
 import { InvalidZodSchemaError, RouterError } from "./error.js"
 import type { EndpointConfig, ContextSearchParams, ContentType } from "./types.js"
-import { SerializeOptions } from "cookie"
 
 /**
  * @experimental