+++ src/types.ts
@@ -1,7 +1,6 @@
 import { type ZodObject, z } from "zod"
 import { RouterError } from "./error.js"
 import { HeadersBuilder } from "./headers.js"
-import { type IncomingHttpHeaders } from "http"
 
 /**
  * Route pattern must start with a slash and can contain parameters prefixed with a colon.
@@ -312,5 +311,77 @@ export interface ClientOptions {
     /**
      * Default headers to include in every request made by the client.
      */
-    headers?: IncomingHttpHeaders
+    headers?: Headers | (() => Headers | Promise<Headers>)
+}
+
+
+export interface Headers extends Record<string, number | string | string[] | undefined> {
+    accept?: string | undefined;
+    "accept-encoding"?: string | undefined;
+    "accept-language"?: string | undefined;
+    "accept-patch"?: string | undefined;
+    "accept-ranges"?: string | undefined;
+    "access-control-allow-credentials"?: string | undefined;
+    "access-control-allow-headers"?: string | undefined;
+    "access-control-allow-methods"?: string | undefined;
+    "access-control-allow-origin"?: string | undefined;
+    "access-control-expose-headers"?: string | undefined;
+    "access-control-max-age"?: string | undefined;
+    "access-control-request-headers"?: string | undefined;
+    "access-control-request-method"?: string | undefined;
+    age?: string | undefined;
+    allow?: string | undefined;
+    "alt-svc"?: string | undefined;
+    authorization?: string | undefined;
+    "cache-control"?: string | undefined;
+    connection?: string | undefined;
+    "content-disposition"?: string | undefined;
+    "content-encoding"?: string | undefined;
+    "content-language"?: string | undefined;
+    "content-length"?: string | undefined;
+    "content-location"?: string | undefined;
+    "content-range"?: string | undefined;
+    "content-type"?: string | undefined;
+    cookie?: string | undefined;
+    date?: string | undefined;
+    etag?: string | undefined;
+    expect?: string | undefined;
+    expires?: string | undefined;
+    forwarded?: string | undefined;
+    from?: string | undefined;
+    host?: string | undefined;
+    "if-match"?: string | undefined;
+    "if-modified-since"?: string | undefined;
+    "if-none-match"?: string | undefined;
+    "if-unmodified-since"?: string | undefined;
+    "last-modified"?: string | undefined;
+    location?: string | undefined;
+    origin?: string | undefined;
+    pragma?: string | undefined;
+    "proxy-authenticate"?: string | undefined;
+    "proxy-authorization"?: string | undefined;
+    "public-key-pins"?: string | undefined;
+    range?: string | undefined;
+    referer?: string | undefined;
+    "retry-after"?: string | undefined;
+    "sec-fetch-site"?: string | undefined;
+    "sec-fetch-mode"?: string | undefined;
+    "sec-fetch-user"?: string | undefined;
+    "sec-fetch-dest"?: string | undefined;
+    "sec-websocket-accept"?: string | undefined;
+    "sec-websocket-extensions"?: string | undefined;
+    "sec-websocket-key"?: string | undefined;
+    "sec-websocket-protocol"?: string | undefined;
+    "sec-websocket-version"?: string | undefined;
+    "set-cookie"?: string[] | undefined;
+    "strict-transport-security"?: string | undefined;
+    tk?: string | undefined;
+    trailer?: string | undefined;
+    "transfer-encoding"?: string | undefined;
+    upgrade?: string | undefined;
+    "user-agent"?: string | undefined;
+    vary?: string | undefined;
+    via?: string | undefined;
+    warning?: string | undefined;
+    "www-authenticate"?: string | undefined;
 }